moved benchmarks to separate file
