+++ merkle_util_test.go
@@ -1059,22 +1059,3 @@ func verifyBlockRootAgainstBeaconStateViaHistoricalSummaries(oracleBeaconStateRo
 	}
 	return epgcommon.ValidateProof(oracleBeaconStateRoot, proof, beaconBlockHeaderToVerifyRoot, historicalBlockHeaderIndex)
 }
-
-func BenchmarkComputeBeaconStateRoot(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		_, _ = epp.ComputeBeaconStateRoot(&oracleState)
-	}
-
-}
-
-func BenchmarkComputeBeaconStateTopLevelRoots(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		_, _ = epp.ComputeBeaconStateTopLevelRoots(&spec.VersionedBeaconState{Deneb: &oracleState})
-	}
-}
-
-func BenchmarkComputeValidatorTree(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		_, _ = epp.ComputeValidatorTree(oracleState.Slot, oracleState.Validators)
-	}
-}
